+++ xenon-common/src/main/java/com/vmware/xenon/services/common/authn/AuthenticationRequest.java
@@ -20,4 +20,6 @@ public class AuthenticationRequest {
     }
 
     public AuthenticationRequestType requestType;
+    public Long sessionExpirationSeconds;
+
 }